bug fix, only try to delete local files

+++ tasks/lib/encode.js
@@ -91,10 +91,11 @@ exports.init = function(grunt) {
           complete();
         } else {
           // process it and put it into the cache
-          var loc = img;
+          var loc = img,
+            is_local_file = !rData.test(img) && !rExternal.test(img);
 
           // Resolve the image path relative to the CSS file
-          if(!rData.test(img) && !rExternal.test(img)) {
+          if(is_local_file) {
             // local file system.. fix up the path
             loc = img.charAt(0) === "/" ?
               (opts.baseDir || "") + loc :
@@ -111,7 +112,7 @@ exports.init = function(grunt) {
             if (err == null) {
               result += "url(" + resp + ")";
 
-              if(deleteAfterEncoding) {
+              if(deleteAfterEncoding && is_local_file) {
                 grunt.log.writeln("deleting " + loc);
                 fs.unlinkSync(loc);
               }
